requirements.txt
pyttsx3
SpeechRecognition
pywhatkit
pyjokes
wikipedia-api
pyaudio
googletrans
pyowm